api/modelApi/requirements.txt
fastapi
tensorflow
pillow
opencv
python-multipart